Split git helpers into their own files to satisfy the file length budget
FileExplorerStore.swift and cmux_open.swift were both at their Swift
file-length budget ceiling. Move the Git Status block (GitFileStatus +
GitStatusProvider) to Sources/GitStatusProvider.swift and the diff CLI
git subprocess helpers to CLI/CMUXCLI+GitProcess.swift. Mechanical moves,
no behavior change; the moved CLI helpers drop 'private' since they are
now called across files within the same target.
